feat: 컬렉션 검색 Service layer 개발 중 - QueryDSL 적용 예정

+++ src/main/java/com/interview_master/ui/CollectionSearchController.java
@@ -16,7 +16,7 @@ public class CollectionSearchController {
 
     @QueryMapping
     public CollectionSearchResult searchCollections(
-            @Argument List<Integer> categoryIds, @Argument List<String> keywords,
+            @Argument List<Long> categoryIds, @Argument List<String> keywords,
             @Argument Integer maxCorrectRate, @ContextValue(required = false) Long userId) {
 
         // 비회원 접근 -> PUBLIC인 컬렉션만 리턴
